+++ emgapi/urls.py
@@ -125,7 +125,7 @@ urlpatterns += [
 
     url(
         (r'^pipeline-tools/(?P<tool_name>[\w+]+)/'
-         '(?P<version>[a-zA-Z0-9\-\.]+)$'),
+         r'(?P<version>[a-zA-Z0-9\-\.]+)$'),
         views.PipelineToolAPIView.as_view(),
         name='pipeline-tools-detail'
     ),
