+++ emgapimetadata/management/commands/import_summary.py
@@ -4,48 +4,19 @@
 import os
 import csv
 
-from django.core.management.base import BaseCommand
-
 from emgapi import models as emg_models
 
+from ..lib import EMGBaseCommand
 
-class Command(BaseCommand):
 
-    obj_list = list()
-    rootpath = None
-    accession = None
+class Command(EMGBaseCommand):
 
-    def add_arguments(self, parser):
-        parser.add_argument('accession', type=str)
-        parser.add_argument('rootpath', type=str)
-
-    def handle(self, *args, **options):
-        self.rootpath = options.get('rootpath', None)
-        self.accession = options.get('accession', None)
-        self.find_accession(options)
-        self.populate_from_accession()
-
-    def find_accession(self, options):
-        if self.accession:
-            self.obj_list = emg_models.AnalysisJob.objects \
-                .filter(sample__study__accession=self.accession).available()
-            if len(self.obj_list) < 1:
-                self.obj_list = emg_models.AnalysisJob.objects \
-                    .filter(sample__accession=self.accession)
-            if len(self.obj_list) < 1:
-                self.obj_list = emg_models.AnalysisJob.objects \
-                        .filter(accession=self.accession)
-            if len(self.obj_list) < 1:
-                raise AttributeError(
-                    "Invalid accession: %s" % self.accession)
-
-    def populate_from_accession(self):
+    def populate_from_accession(self, options):
         for o in self.obj_list:
             self.find_path(o)
 
     def find_path(self, obj):
         res = os.path.join(self.rootpath, obj.result_directory)
-        print(obj.result_directory, obj.accession)
         if os.path.exists(res):
             if os.path.isdir(res):
                 for root, dirs, files in os.walk(res, topdown=False):
