+++ emgapimetadata/urls.py
@@ -24,11 +24,13 @@ from . import views as m_views
 from . import other_views as o_views
 
 app_name = "emgapimetadata"
-urlpatterns = [
+urlpatterns = []
+
+urlpatterns += [
 
     url(
         (r'^runs/(?P<accession>[a-zA-Z0-9_]+)/(?P<release_version>[0-9\.]+)/'
-         'annotations$'),
+         r'annotations$'),
         m_views.AnnotationAnalysisAPIView.as_view(),
         name='runs-pipelines-annotations-list'
     ),
